requirements.txt
django==4.2.2
djangorestframework==3.14.0
decouple==0.0.7